Swapping my poll-contextual tweet copy for Kurt's
Or am I swapping Kurt's *in for* mine? Let's turn to the New Oxford
American Dictionary for some related usage notes:

> Traditionally, the verb substitute is followed by for and means ‘put
(someone or something) in place of another,’ as in she substituted the
fake vase for the real one. From the late 17th century, substitute has
also been used to mean ‘replace (someone or something) with someone or
something else,’ as in she substituted the real vase with the fake one.
This can be confusing, since the two sentences shown above mean the
same thing, yet the object of the verb and the object of the
preposition have swapped positions. Despite the potential confusion,
the second, newer use is well established, especially in some
scientific contexts and in sports ( the top scorer was substituted with
almost half an hour still to play), and is now generally regarded as
part of standard English.